Updates See DESC
Updated TODO List...not all...
Updated Skills Animations
Updated SkillCasting following Client Logics...
Sync some test Functions like UG_CHAR_TRANSFORM_REQ and SendUseItem

+++ Server/GameServer/QueryClass.cpp
@@ -205,7 +205,7 @@ bool	QueryClass::UpdateSPPoint(RwUInt32 CharID, RwUInt32 point)
 bool	QueryClass::InsertRemoveQuickSlot(TBLIDX SkillID, BYTE SlotID, int CharID)
  {
  	CGameServer * app = (CGameServer*)NtlSfxGetApp();
-	std::string query = "UPDATE quickslot SET slotId_" +std::to_string((long double)(SlotID));
+ 	std::string query = "UPDATE quickslot SET slotId_" +std::to_string(SlotID);
  	app->db->prepare(query+" = ? WHERE charId=?");
  	app->db->setInt(1, SkillID);
  	app->db->setInt(2, CharID);
